fixes compilation errors
- fixes #8
